requirements.txt
aiogram

python-dotenv
uvicorn
aiohttp
